refactor undo modal into its own component
